requirements.txt
fastapi
uvicorn[standard]
anthropic
pydantic
sqlalchemy[asyncio]
asyncpg
aiosqlite
httpx
redis
python-dotenv
